Change where function redirect are
